Move entity models to entity package, add password and email to organization model
